Fixed holiday checker and applied holiday checker for patient and healthprof
