+++ app/src/main/res/layout/activity_login.xml
@@ -31,7 +31,7 @@
                 android:layout_gravity="center"
                 android:layout_marginTop="30dp"
                 android:layout_marginBottom="10dp"
-                android:text="WELCOME TO DOCTRACT!"
+                android:text="WELCOME TO DOCTRACK!"
                 android:textColor="#FFFFFF"
                 android:textSize="25sp"
                 android:textStyle="bold" />
